GPUHandle: int id for synchronously sending objects between threads
The motivation for this is primarily for WebAssembly applications.
In native code, the natural way to share things between threads is
simply to use the object from multiple threads. However, due to the
asynchronous nature of postMessage, this becomes impossible to do: even
if the app provides a signal that it wants to transfer between threads,
it must wait for a callback before it can use the object.

In an application worker which periodically releases control to the
browser, it's *technically* possible to:

- send requests to the owning thread to postMessage the appropriate
  objects, and
- buffer up all of the submitted work using custom ids, wait for all of
  the objects to arrive via onmessage, and then replay the work.

However, this requires the application to release control frequently,
and incurs a significant delay in processing of commands.

The removal of [Serializable] is optional.