createUploadBuffer
Add createUploadBuffer, which has the explicit purpose of creating
temporary storage for uploads. It is essentially identical to
createBufferMapped, but less flexible: there is no usage field. The
usage can only be MAP_WRITE | COPY_SRC.

The expectation with GPUUploadBuffer is that it would generally be
allocated as CPU memory (e.g. IPC shared memory) - or, if available, a
suballocation in some pre-allocated CPU-coherent memory (e.g.
HOST_COHERENT mapped with dmabuf).